// ClientApp/src/app/models/outlookview.model.ts
export class Radio {
    check: boolean;
    itemsCommerce: string;
}

export class ActiveProduct {
    commerce: string;
    index: number;
    keword: string;
    Product: any[];
}

export class ItemsCommerce {
    model: string;
    brand: string;
    check: boolean;
    commerce: string;
    prices: string;
    urlDetail: string;
    hidenspan: boolean;
    idcheck: string;
}

export class ModelBrand {
    model: string;
    brand: string;
    commerceName: string;
}

export class ActiveComerce {
   commerce: string;
   isActive: boolean;
}